+++ service/controller/rbac/resource/externalresources/resource.go
@@ -6,7 +6,7 @@ package externalresources
 import (
 	"reflect"
 
-	"github.com/giantswarm/k8sclient/v5/pkg/k8sclient"
+	"github.com/giantswarm/k8sclient/v7/pkg/k8sclient"
 	"github.com/giantswarm/microerror"
 	"github.com/giantswarm/micrologger"
 	rbacv1 "k8s.io/api/rbac/v1"